Cleanup and stuff for the cart display page
- Gather stuff for each functional unit into its own file
- DRY up the StyledCard (should probably rename this but it's ok for now)
- Update the index page to use the new components and such

We still don't post things but the APIs aren't exposed properly, so have to fix that first (in the Django app).
