+++ src/page-components/Card/StyledCard.tsx
@@ -1,0 +1,8 @@
+import { styled } from "@mitodl/smoot-design";
+import { Card } from "./Card";
+
+const StyledCard = styled(Card)`
+padding: 32px;
+`;
+
+export default StyledCard;
